deleted server info, fixing permissions issues for clear command

+++ commands/clear.js
@@ -10,6 +10,8 @@ module.exports = {
     if(args[0] > 100) return message.reply(config.messages['too-much-clear-error']);
     if(args[0] < 1) return message.reply(config.messages['not-enough-clear-error']);
 
+    if(!message.member.hasPermission('MANAGE_MESSAGES')) return message.reply(config.messages['no-permission-error']);
+
     await message.channel.messages.fetch({limit: args[0]}).then(messages => {
       message.channel.bulkDelete(messages);
     })
